+++ srcs/en/swingy/equipement/armor/Armor.java
@@ -3,15 +3,15 @@ package en.swingy.equipement.armor;
 import en.swingy.equipement.Equipement;
 
 public class Armor extends Equipement {
-	protected final int defense;
+	protected final float defense;
 
-	public Armor(String p_name, int p_defense) {
+	public Armor(String p_name, float p_defense) {
 		super(p_name);
 		this.defense = p_defense;
 	}
 
 	@Override
-	public int getDefense() {
+	public float getDefense() {
 		return this.defense;
 	}
 }
