+++ srcs/en/swingy/equipement/weapon/Weapon.java
@@ -3,7 +3,7 @@ package en.swingy.equipement.weapon;
 import en.swingy.equipement.Equipement;
 
 public class Weapon extends Equipement {
-	protected final int attack;
+	protected final float attack;
 
 	public Weapon(String p_name, int p_attack) {
 		super(p_name);
@@ -11,7 +11,7 @@ public class Weapon extends Equipement {
 	}
 
 	@Override
-	public int getAttack() {
+	public float getAttack() {
 		return this.attack;
 	}
 }
